Fix all clippy warnings, and deny them in CI
* Fix all clippy lints

Also deny clippy warnings in CI

* Changelog

* Fix lints left out

* Changelog: specify that clippy denies warnings in CI

* Update PR template with clippy recommendations

+++ hal/src/rtc.rs
@@ -455,7 +455,7 @@ impl TimerParams {
                                               * (rust-lang/rust#51999) */
         };
 
-        let cycles: u32 = ticks / divider_value as u32;
+        let cycles: u32 = ticks / divider_value;
 
         TimerParams { divider, cycles }
     }
